refactor(capture): align list with screen layout
